requirements.txt
beautifulsoup4==4.13.3
pandas==2.2.3
protobuf==5.29.0
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.8.3
openpyxl==3.1.5
